requirements (3).txt
streamlit
pandas
scikit-learn
google-generativeai
python-dotenv
imbalanced-learn
PyMuPDF
fpdf
Pillow